+++ tests/all/utf_8_characters.rs
@@ -3,7 +3,10 @@ use pretty_assertions::assert_eq;
 use comfy_table::*;
 
 #[test]
-fn simple_emoji_table() {
+/// UTF-8 symbols that are longer than a single character are properly handled.
+/// This means, that comfy-table detects that they're longer than 1 character and styles/arranges
+/// the table accordingly.
+fn multi_character_utf8_symbols() {
     let mut table = Table::new();
     table
         .set_header(&vec!["Header1", "Header2", "Header3"])
@@ -29,5 +32,28 @@ fn simple_emoji_table() {
 |                      | add some             |                        |
 |                      | multi line stuff     |                        |
 +----------------------+----------------------+------------------------+";
+    assert_eq!("\n".to_string() + &table.to_string(), expected);
+}
+
+#[test]
+///
+fn multi_character_utf8_word_splitting() {
+    let mut table = Table::new();
+    table
+        .set_table_width(8)
+        .set_content_arrangement(ContentArrangement::Dynamic)
+        .set_header(&vec!["test"])
+        .add_row(&vec!["abc✅def"]);
+
+    println!("{}", table.to_string());
+    let expected = "
++------+
+| test |
++======+
+| abc  |
+| ✅de |
+| f    |
++------+";
+    println!("{}", expected);
     assert_eq!("\n".to_string() + &table.to_string(), expected);
 }